# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel 11 REST API Sample Project",
        "description": "Explore the capabilities of Laravel 11 by building a RESTful API. This sample project demonstrates best practices and key features such as the repository pattern, PHPUnit testing, and Swagger documentation.",
        "contact": {
            "name": "Jomar Sulit",
            "email": "sulitjomar@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/actividades": {
            "get": {
                "tags": [
                    "Actividad"
                ],
                "summary": "Get all actividades",
                "operationId": "c801ac63fc582a8e85d03083d61ff9e3",
                "responses": {
                    "200": {
                        "description": "List of actividades"
                    }
                }
            }
        },
        "/actividades/{id}": {
            "get": {
                "tags": [
                    "Actividad"
                ],
                "summary": "Get a single actividad",
                "operationId": "03d25023f748fe1d7f1b69896798d13c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad details"
                    },
                    "404": {
                        "description": "Actividad not found"
                    }
                }
            }
        },
        "/actividades/{actividadId}/categorias": {
            "get": {
                "tags": [
                    "Actividad"
                ],
                "summary": "Get all categories for a specific activity",
                "operationId": "ef26911c1e011c9cdfafe295103969d6",
                "parameters": [
                    {
                        "name": "actividadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories"
                    }
                }
            }
        },
        "/actividades/{actividadId}/alumnos": {
            "get": {
                "tags": [
                    "Actividad"
                ],
                "summary": "Get all students for a specific activity",
                "operationId": "8902e5c64be4b3fcf001bcd05ff0455f",
                "parameters": [
                    {
                        "name": "actividadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of students"
                    }
                }
            }
        },
        "/actividades/{actividadId}/categorias/{categoriaId}": {
            "post": {
                "tags": [
                    "Actividad"
                ],
                "summary": "Associate a category with an activity",
                "operationId": "6f2234a2a3c86b0ed286b69d0d58759f",
                "parameters": [
                    {
                        "name": "actividadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category associated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Actividad"
                ],
                "summary": "Remove association between a category and an activity",
                "operationId": "51059019c9e87f8e7d1a089c70e1c873",
                "parameters": [
                    {
                        "name": "actividadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category disassociated successfully"
                    }
                }
            }
        },
        "/alumnos": {
            "get": {
                "tags": [
                    "Alumno"
                ],
                "summary": "Get all alumnos",
                "operationId": "70304f65ed5181ea6388593e07279822",
                "responses": {
                    "200": {
                        "description": "List of alumnos"
                    }
                }
            }
        },
        "/alumnos/{id}": {
            "get": {
                "tags": [
                    "Alumno"
                ],
                "summary": "Get a single alumno",
                "operationId": "92f23a1367b13c816ca8b7aee397a3f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alumno details"
                    },
                    "404": {
                        "description": "Alumno not found"
                    }
                }
            }
        },
        "/alumnos/{alumnoId}/actividades": {
            "get": {
                "tags": [
                    "Alumno"
                ],
                "summary": "Get all activities for a specific student",
                "operationId": "79cfb23062287525fa31907472f77cab",
                "parameters": [
                    {
                        "name": "alumnoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of activities"
                    }
                }
            }
        },
        "/alumnos/{alumnoId}/actividades/{actividadId}": {
            "post": {
                "tags": [
                    "Alumno"
                ],
                "summary": "Associate an activity with a student",
                "operationId": "3e2c71801ecf3dd21e94f21d2e62a4e8",
                "parameters": [
                    {
                        "name": "alumnoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "actividadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity associated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alumno"
                ],
                "summary": "Remove association between an activity and a student",
                "operationId": "e8a8c59f9689896e00181e69ae945995",
                "parameters": [
                    {
                        "name": "alumnoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "actividadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity disassociated successfully"
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of books",
                "operationId": "28b9632495ac5f6f66b2e52001d510e6",
                "responses": {
                    "200": {
                        "description": "A list of books"
                    }
                }
            }
        },
        "/categorias": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Get all categorias",
                "operationId": "66f747250d1c2a1a99b78b5ffc76ad86",
                "responses": {
                    "200": {
                        "description": "List of categorias"
                    }
                }
            }
        },
        "/categorias/{id}": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Get a single categoria",
                "operationId": "c850dd742053be302993ad6e3c2f9f2b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria details"
                    },
                    "404": {
                        "description": "Categoria not found"
                    }
                }
            }
        },
        "/categorias/{categoriaId}/actividades": {
            "get": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Get all activities for a specific category",
                "operationId": "f7f11850302776d8f2d116e3c99acfb2",
                "parameters": [
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of activities"
                    }
                }
            }
        },
        "/categorias/{categoriaId}/actividades/{actividadId}": {
            "post": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Associate an activity with a category",
                "operationId": "4c73efa1327488171df79ad768d1bf08",
                "parameters": [
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "actividadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity associated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categoria"
                ],
                "summary": "Remove association between an activity and a category",
                "operationId": "82c5d105ea488c72d065f2febca937bf",
                "parameters": [
                    {
                        "name": "categoriaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "actividadId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity disassociated successfully"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a list of products",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of products"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No Products Available"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a new product",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "details"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Product Name"
                                    },
                                    "details": {
                                        "type": "string",
                                        "example": "Product Details"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Display a specific product",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a specific product",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the product",
                                        "type": "string"
                                    },
                                    "details": {
                                        "description": "Details of the product",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a specific product",
                "operationId": "fe34b364a3f192305598b65d04d06b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "Product ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "details": {
                        "description": "Product details",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Actividad",
            "description": "API Endpoints of Actividad"
        },
        {
            "name": "Alumno",
            "description": "API Endpoints of Alumno"
        },
        {
            "name": "Books",
            "description": "API Endpoints of Products"
        },
        {
            "name": "Categoria",
            "description": "API Endpoints of Categoria"
        },
        {
            "name": "Products",
            "description": "API Endpoints of Products"
        }
    ]
}